+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Changed
 - Slider in shared regions discovery to filter nodes by pvalue (filter by pval from 1 to 10^-10) or size (filter from 1 to 1000)
+- Use docker compose instead of docker to launch neo4j db
 
 ### Fixed
 - Node and edge information display